Clean up rehydration code into a utility module.
